api_service/requirements.txt
fastapi
uvicorn
redis
rq
httpx
prometheus_client
fastapi-cache2
